tests/aoc-2020-csharp-tests/Day09Tests.cs
using aoc_2020_csharp.Day09;

namespace aoc_2020_csharp_tests;

public class Day09Tests
{
    [Test]
    public void Part1_Example_ReturnsCorrectAnswer()
    {
        var input = new[] { 35L, 20, 15, 25, 47, 40, 62, 55, 65, 95, 102, 117, 150, 182, 127, 219, 299, 277, 309, 576 };
        var expected = 127L;
        var actual = Day09.Solve1(input, 5);
        actual.Should().Be(expected);
    }

    [Test]
    public void Part1_ReturnsCorrectAnswer()
    {
        var expected = 88311122L;
        var actual = Day09.Part1();
        actual.Should().Be(expected);
    }

    [Test]
    public void Part2_Example_ReturnsCorrectAnswer()
    {
        var input = new[] { 35L, 20, 15, 25, 47, 40, 62, 55, 65, 95, 102, 117, 150, 182, 127, 219, 299, 277, 309, 576 };
        var expected = 62L;
        var actual = Day09.Solve2(input, 5);
        actual.Should().Be(expected);
    }

    [Test]
    public void Part2_ReturnsCorrectAnswer()
    {
        var expected = 13549369;
        var actual = Day09.Part2();
        actual.Should().Be(expected);
    }
}
